Add support for skipping user input in tests when NO_INTERACTIVE_MODE is defined

+++ tests/test_coco.cpp
@@ -12,10 +12,12 @@
 #endif
 #ifdef BUILD_SERVER
 #include "coco_server.hpp"
+#endif
+#if defined(BUILD_SERVER) || defined(NO_INTERACTIVE_MODE)
 #include <thread>
 #endif
 
-int main()
+int main([[maybe_unused]] int argc, [[maybe_unused]] char *argv[])
 {
 #ifdef BUILD_MONGODB
     mongocxx::instance inst{}; // This should be done only once.
@@ -48,10 +50,22 @@ int main()
     auto &s_itm = cc.create_item(s_tp);
     cc.set_value(s_itm, json::json{{"parent", ch_itm.get_id().c_str()}});
 
-    std::string user_input;
-    std::cin >> user_input;
-    if (user_input == "d")
-        db.drop();
+#ifndef NO_INTERACTIVE_MODE
+    std::this_thread::sleep_for(std::chrono::seconds(1));
+    db.drop();
+#else
+    bool skip_user_input = false;
+    if (argc > 1 && std::string(argv[1]) == "--skip-input")
+        skip_user_input = true;
+
+    if (!skip_user_input)
+    {
+        std::string user_input;
+        std::cin >> user_input;
+        if (user_input == "d")
+            db.drop();
+    }
+#endif
 
 #ifdef BUILD_SERVER
     srv.stop();
